+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - change: delivery format to just "format" in ui
 - fix: jsonjsdb with minimal data
 - fix: HTTP server compatibility with Windows network drives (SMB shares)
+- fix: remove z-index from frequency value styles
 - remove: md extension from LICENSE file
 
 ## 0.16.0 (2025-12-14)