+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 - add: initial data template files for jsonjsdb, static-make, and update-app
 - add: alternative db for testing
+- change: delivery format to just "format" in ui
 - fix: jsonjsdb with minimal data
 - fix: HTTP server compatibility with Windows network drives (SMB shares)
+- fix: remove z-index from frequency value styles
 - remove: md extension from LICENSE file
 
 ## 0.16.0 (2025-12-14)